fix impl POST+COMMENT

+++ src/main/java/org/diplom/blog/dto/mapper/TagMapper.java
@@ -5,7 +5,7 @@ import org.diplom.blog.dto.TagDto;
 import org.diplom.blog.model.Tag;
 
 public class TagMapper {
-    public static TagDto toTagDto(EntityCount<Tag> tagCount, int commonCount) {
+    public static TagDto toTagDto(EntityCount<Tag> tagCount, Long commonCount) {
         return new TagDto(tagCount.getEntity().getName(),
                 (double) tagCount.getCountRecord() / commonCount);
     }
